fix: GPU inference pipeline - embed_tokens/dtype/transform bugs, add vis script
- Fix embed_tokens size mismatch (checkpoint vocab=4100, was loading BPE dict -> 54360)
- Remove .half() hardcodings in swin.py:290 and unify_transformer.py:1336 for CPU/GPU fp32 compatibility
- Fix reverse_transform order (fliplr before rot90, not after)
- Fix transpose axis from (2,1,0) to (1,2,0) for correct orientation
- Add run_inference_vis.py: per-slice Dice CSV + Top/Bottom PNG panels
- GPU (RTX 3090) enabled, 275 slices in ~9 min
- Known issue: Dice=0 due to coordinate scaling (train /512 vs inference /256)

+++ models/polyformer/swin.py
@@ -287,7 +287,7 @@ class WindowAttention(nn.Module):
         else:
             attn = self.softmax(attn)
 
-        attn = self.attn_drop(attn).half()
+        attn = self.attn_drop(attn)
 
         x = (attn @ v).transpose(1, 2).reshape(B_, N, C)
         x = self.proj(x)
